Implement new type system. (#297)
Closes #9, #22, #50, #298, #29

+++ bundles/org.eclipse.mita.base/src/org/eclipse/mita/base/expressions/terminals/HexValueConverter.java
@@ -20,30 +20,30 @@ import org.eclipse.xtext.util.Strings;
  * @author axel terfloth
  *
  */
-public class HexValueConverter extends AbstractLexerBasedConverter<Integer> {
+public class HexValueConverter extends AbstractLexerBasedConverter<Long> {
 
 	public static final String HEX_PREFIX = "0x";
 	
-	public Integer toValue(String string, INode node) {
+	public Long toValue(String string, INode node) {
 
 		if (Strings.isEmpty(string))
 			throw new ValueConverterException("Couldn't convert empty string to number.", node, null);
 
 		try {
 			// perform the conversion with string index 2 since the prefix is always '0x'
-			return Integer.parseInt(string.substring(2), 16);
+			return Long.parseLong(string.substring(2), 16);
 		} catch ( NumberFormatException e ) {
 			throw new ValueConverterException("Couldn't convert '" + string + "' to number.", node, null);
 		}
 	}
 
 	@Override
-	protected String toEscapedString(Integer value) {
+	protected String toEscapedString(Long value) {
 		if (value < 0) { 
-			return "-" + HEX_PREFIX + Integer.toString( value * -1, 16).toUpperCase();
+			return "-" + HEX_PREFIX + Long.toString( value * -1, 16).toUpperCase();
 
 		}
-		return HEX_PREFIX + Integer.toString(value, 16).toUpperCase();
+		return HEX_PREFIX + Long.toString(value, 16).toUpperCase();
 	}
